requirements.txt
# This file is expected by Heroku.

#-r requirements/production.txt

amqp==2.4.2
argon2-cffi==19.1.0
arrow==0.13.1
Babel==2.6.0
billiard==3.5.0.5
binaryornot==0.4.4
celery==4.2.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
defusedxml==0.5.0
Django==2.0.13
django-allauth==0.38.0
django-appconf==1.0.3
django-compressor==2.2
django-coverage-plugin==1.6.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-environ==0.4.5
django-extensions==2.1.6
django-model-utils==3.1.2
django-redis==4.10.0
django-rest-swagger==2.2.0
djangorestframework==3.9.1
flower==0.9.2
future==0.17.1
idna==2.8
itypes==1.1.0
Jinja2==2.10
jinja2-time==0.2.0
kombu==4.4.0
MarkupSafe==1.1.1
numpy==1.16.2
oauthlib==3.0.1
openapi-codec==1.3.2
pandas==0.24.2
Pillow==5.4.1
poyo==0.4.2
pycparser==2.19
python-dateutil==2.8.0
python-slugify==2.0.1
python3-openid==3.1.0
pytz==2018.9
rcssmin==1.0.6
redis==2.10.6
requests==2.21.0
requests-oauthlib==1.2.0
rjsmin==1.0.12
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
tornado==6.0.1
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
whichcraft==0.5.2
whitenoise==4.1.2
